skype/consumer/requirements.txt
skpy
pika
python-dotenv
